Refactor MdeaNode
Instead of a single node struct where the data is an union field, there
are now separate structs for nodes with different payloads so they're
more tightly packed.

Node type is determined by a pointer to a MdeaNodeType struct that
holds pointers to key functions for handling the data associated with
the node.