Implement mdea_parser and mdea_emitter
struct mdea_parser and its related functions transform an input element
that can be either a file or a wide-character string into a token
stream that can be parsed into a node.

struct mdea_emitter and its related functions allow for the opposite
process, formatting a node into a file or string.